Integration tests with hurl
