+++ integration-tests/gourmet/go.mod
@@ -1,0 +1,3 @@
+module integration
+
+go 1.23.1
