+++ integration-tests/gourmet/main.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"fmt"
+	"os/exec"
+)
+
+func main() {
+	output, _ := exec.Command("go run").Output()
+	fmt.Println(output)
+}
